Use the ubuntu user everywhere as it has ID 1000
This is a hack to support the current build script as it requires the
user ID inside the container to match the external user which, for my
personal laptop, is 1000. This affects the bind-mount for the SSH Agent
socket where the agent checks the user id on the remote end and rejects
invalid user IDs.