Attempt to use pristine-tar and/or Debian archive
As the latest upstream release is not in Ubuntu yet, we need to either
download it from Debian's servers or, preferrably, get it from the
pristine-tar branch.

Also, fixed an issue with apt-get build-dep. In it's previous
invocation, it was using the current package in the archive to locate
dependencies, not the current Git checkout.